Created MenuComponent
