+++ src/components/MenuComponent/styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Menu = styled.div`
+    padding: 0px 20px;
+
+    .fa-bars {
+        font-size: 20px;
+    }
+`;
